requirements.txt
chardet==5.2.0
contourpy==1.3.0
cycler==0.12.1
fonttools==4.54.1
kiwisolver==1.4.7
matplotlib==3.9.2
packaging==24.1
pandas==2.2.3
pillow==11.0.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytz==2024.2
reportlab==4.2.5
seaborn==0.13.2
six==1.16.0
tzdata==2024.2
tensorflow-macos
tensorflow-metal
numpy<2.0
pyyaml
